fixes a bug in package handling
This is mostly some handling for https://tickets.puppetlabs.com/browse/PUP-1073
The workaround moves from the puppet manigest to the pecl provider.

It is now possible to specify pecl packages with a 'pecl-' prefix to
avoid declaring multiple packages with the same name. This still is an
ugly hack but the least ugly I can think of currently.