Rely more on puppetlabs-apt
This refactors the code to rely more on puppetlabs-apt features. It
passes in the key and lets apt::source define apt::key for us.

Puppetlabs-apt also knows better when apt-transport-https is actually
needed. On the most recent Debian/Ubuntu releases it doesn't even exist
anymore since apt handles it out of the box.

It also removes the absolute resource notation which is no longer
needed.