php-fpm: Make service reload/restart configurable
By default, this module uses a reload for the php-fpm service. This is
nice because it ensures that there is no downtime. But this is also bad
because certain configuration options, like unix socket owner/group will
only be changed during a restart.

Fixes: https://github.com/voxpupuli/puppet-php/issues/596
Replaces: https://github.com/voxpupuli/puppet-php/pull/597