Allow empty string for setting value
Some PHP features are enabled by default and can be disabled by using an
empty string as a setting value, e.g. user_ini.filename.  The code
currently require a non-empty string, making it impossible to disable
such a feature.

Allow an empty string as the value of a setting.